finish

+++ src/item.rs
@@ -6,8 +6,8 @@
     pub struct item {
         pub name: String,
         pub time_mult: f32,    // for skateboard
-        pub hunger_effect: i8, // e.g. snack
-        pub health_effect: i8, // e.g. someone poisoined ur snack :(
+        pub hunger_effect: i16, // e.g. snack
+        pub health_effect: i16, // e.g. someone poisoined ur snack :(
     }
 
     // impl Default for item {
@@ -34,7 +34,7 @@
         pub fn snack() -> item {
             // let number: usize = thread_rng().gen_range(0..10); -- this is for rng but it wasnt workign at the time of me writing this lmao
             let number = 2;
-            let healthEffect: i8;
+            let healthEffect: i16;
             //Leaving it up to RNG if it's experied or not
             if number > 5 {
                 healthEffect = 10;
